+++ CloudNotes/CloudNotes/Info.plist
@@ -9,7 +9,7 @@
 			<string></string>
 			<key>CFBundleURLSchemes</key>
 			<array>
-				<string>db-cw6gr31z68zgvkm</string>
+				<string>db-bg6xjvljgmkc6ui</string>
 			</array>
 		</dict>
 	</array>
